requirements.txt
Brotli==1.0.9
inflate64==0.3.1
multivolumefile==0.2.3
psutil==5.9.4
py7zr==0.20.4
pybcj==1.0.1
pycryptodomex==3.17
pyppmd==1.0.0
pyzstd==0.15.4
texttable==1.6.7
tqdm==4.64.1
tree-sitter==0.20.1
torch
dgl
pytorch_lightning
pytorch_metrics